+++ packages/web-core/__tests__/components/history.spec.ts
@@ -25,7 +25,7 @@ import {
   Routing,
 } from '../../src/entities/inworld_packet.entity';
 import { ExtendedHistoryItem } from '../data_structures';
-import { createCharacter, getPacketId, user } from '../helpers';
+import { createCharacter, getPacketId, SCENE, user } from '../helpers';
 
 const characters = [createCharacter(), createCharacter()];
 const packetId = getPacketId();
@@ -89,6 +89,22 @@ const interactionEndPacket = new InworldPacket({
     type: InworlControlType.INTERACTION_END,
   },
 });
+const sceneChangePacketRequest = new InworldPacket({
+  packetId,
+  routing,
+  date,
+  type: InworldPacketType.SCENE_MUTATION_RESPONSE,
+  sceneMutation: { name: v4() },
+});
+const sceneChangePacketResponse = new InworldPacket({
+  packetId,
+  routing,
+  date,
+  type: InworldPacketType.SCENE_MUTATION_RESPONSE,
+  sceneMutation: {
+    addedCharacters: [characters[0]],
+  },
+});
 const incomingTextPacket = new InworldPacket({
   packetId: {
     ...getPacketId(),
@@ -114,6 +130,7 @@ const createHistoryWithPacket = (
   const history = new InworldHistory({
     ...(extension && { extension }),
     user,
+    scene: SCENE,
   });
 
   history.addOrUpdate({ characters, grpcAudioPlayer, packet });
@@ -122,7 +139,7 @@ const createHistoryWithPacket = (
 };
 
 test('should be empty by default', () => {
-  const history = new InworldHistory();
+  const history = new InworldHistory({ scene: SCENE });
 
   expect(history.get().length).toEqual(0);
 });
@@ -328,7 +345,7 @@ describe('text', () => {
 
   describe('transcript', () => {
     test('should return empty transcript for empty history', () => {
-      const history = new InworldHistory();
+      const history = new InworldHistory({ scene: SCENE });
 
       const transcript = history.getTranscript();
 
@@ -407,7 +424,7 @@ describe('text', () => {
           packet: secondtPacket,
         });
 
-        const expected = `${characters[0].displayName}: ${firstPacket.text.text}${secondtPacket.text.text}`;
+        const expected = `${characters[0].displayName}: ${firstPacket.text.text} ${secondtPacket.text.text}`;
         const transcript = history.getTranscript();
         expect(transcript).toEqual(expected);
       });
@@ -442,7 +459,22 @@ describe('text', () => {
           packet: textPacket,
         });
 
-        const expected = `User: (${emotionsPacket.emotions.behavior.code}) ${textPacket.text.text}`;
+        const expected = `User: ${textPacket.text.text}`;
+        const transcript = history.getTranscript();
+
+        expect(transcript).toEqual(expected);
+      });
+
+      test('should return transcript with scene change', () => {
+        const history = createHistoryWithPacket(sceneChangePacketRequest);
+
+        history.addOrUpdate({
+          characters,
+          grpcAudioPlayer,
+          packet: sceneChangePacketResponse,
+        });
+
+        const expected = `>>> Now moving to ${sceneChangePacketRequest.sceneMutation.name}`;
         const transcript = history.getTranscript();
 
         expect(transcript).toEqual(expected);