+++ packages/web-core/src/index.ts
@@ -26,6 +26,7 @@ import {
   HistoryItemActor,
   HistoryItemBase,
   HistoryItemNarratedAction,
+  HistoryItemSceneChange,
   HistoryItemTriggerEvent,
 } from './components/history';
 import { Character } from './entities/character.entity';
@@ -87,6 +88,7 @@ export {
   HistoryItemActor,
   HistoryItemBase,
   HistoryItemNarratedAction,
+  HistoryItemSceneChange,
   HistoryItemTriggerEvent,
   InworldClient,
   InworldConnectionService,